# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChartLogs Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List journal entries
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, journaled, pending]
      responses:
        "200":
          description: List of journal entries with trade data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
  /journal/{tradeId}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get journal entry for a trade
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: upsertJournalEntry
      tags: [journal]
      summary: Create or update a journal entry
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "200":
          description: Journal entry saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  # DASHBOARD
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 1w, 1m, 3m, all]
            default: 1m
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/equity-curve:
    get:
      operationId: getEquityCurve
      tags: [dashboard]
      summary: Get equity curve data points
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 1w, 1m, 3m, all]
            default: 1m
      responses:
        "200":
          description: Equity curve data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquityPoint"
  /dashboard/calendar:
    get:
      operationId: getDashboardCalendar
      tags: [dashboard]
      summary: Get monthly P&L calendar heatmap
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar heatmap
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"
...
  /market/calendar:
    get:
      operationId: getMarketCalendar
      tags: [market]
      summary: Get economic calendar events
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, tomorrow, this_week, all]
            default: today
        - name: impact
          in: query
          schema:
            type: string
            enum: [high, medium, low, all]
            default: all
        - name: currency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Economic calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EconomicEvent"

  # TOOLS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Trade:
      type: object
      required: [id, userId, symbol, type, entryPrice, positionSize, openTime, source, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        symbol:
          type: string
        type:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        positionSize:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        openTime:
          type: string
        closeTime:
          type: ["string", "null"]
        pnl:
          type: ["number", "null"]
        pips:
          type: ["number", "null"]
        rrRatio:
          type: ["number", "null"]
        fees:
          type: ["number", "null"]
        source:
          type: string
          enum: [manual, mt4, mt5, csv]
        tags:
          type: array
          items:
            type: string
        emotion:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string
        outcome:
          type: ["string", "null"]
          enum: [win, loss, breakeven, null]
        createdAt:
          type: string
...
    JournalEntry:
      type: object
      required: [id, tradeId, userId, createdAt]
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        userId:
          type: integer
        notes:
          type: ["string", "null"]
        mood:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string
        trade:
          $ref: "#/components/schemas/Trade"
        createdAt:
          type: string
        updatedAt:
          type: string
    JournalInput:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        mood:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string

    # DASHBOARD
    DashboardStats:
      type: object
      required: [totalPnl, realizedPnl, unrealizedPnl, winRate, totalTrades, openTrades, avgWinner, avgLoser, bestTrade, worstTrade, winStreak, lossStreak, avgRR, profitFactor, expectancy]
      properties:
        totalPnl:
          type: number
        realizedPnl:
          type: number
        unrealizedPnl:
          type: number
        winRate:
          type: number
        totalTrades:
          type: integer
        openTrades:
          type: integer
        avgWinner:
          type: number
        avgLoser:
          type: number
        bestTrade:
          type: number
        worstTrade:
          type: number
        winStreak:
          type: integer
        lossStreak:
          type: integer
        avgRR:
          type: number
        profitFactor:
          type: number
        expectancy:
          type: number
    EquityPoint:
      type: object
      required: [date, equity, pnl]
      properties:
        date:
          type: string
        equity:
          type: number
        pnl:
          type: number
    CalendarDay:
      type: object
      required: [date, pnl, tradeCount]
      properties:
        date:
          type: string
        pnl:
          type: number
        tradeCount:
          type: integer

    # ANALYTICS
...
    EconomicEvent:
      type: object
      required: [id, time, country, currency, impact, event]
      properties:
        id:
          type: string
        time:
          type: string
        country:
          type: string
        currency:
          type: string
        impact:
          type: string
          enum: [high, medium, low]
        event:
          type: string
        actual:
          type: ["string", "null"]
        forecast:
          type: ["string", "null"]
        previous:
          type: ["string", "null"]

    # TOOLS
...