# lib/api-zod/src/generated/types/economicEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChartLogs Trading Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { EconomicEventImpact } from './economicEventImpact';

export interface EconomicEvent {
  id: string;
  time: string;
  country: string;
  currency: string;
  impact: EconomicEventImpact;
  event: string;
  /** @nullable */
  actual?: string | null;
  /** @nullable */
  forecast?: string | null;
  /** @nullable */
  previous?: string | null;
}